requirements.txt
customtkinter==5.2.2
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.4
scipy==1.11.4
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow==2.15.0
torch==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
websocket-client==1.6.4
python-socketio==5.10.0
pillow==10.2.0
seaborn==0.13.0
plotly==5.17.0
streamlit==1.28.0
google-generativeai==0.3.2
groq==0.4.1
ollama==0.2.1
openai==1.6.1
lightgbm==4.1.0
catboost==1.2.2
ta==0.10.2
psutil==5.9.6
rich==13.7.0
